PFW-1561 MMU unload when there are no fan errors
EFCE_FIXED is set to EFCE_OK only when the printer is not active.

This resolves an issue where, while the print job is paused due to the fan error; if you allow the extruder to cool down below 50°C, then the printer will still try to heat up the extruder to do the unload.

Instead of allowing both EFCE_OK and EFCE_FIXED. Only allow unloading if fan_check_error == EFCE_OK