PFW-1561 Use did_pause_print explicitly
print_job_timer is stopped at this point so print_job_timer.paused() always yields false.
For this reason here we only need to check did_pause_print. I think this makes the code
a bit easier to read and gives a reason why we are clearing the flag at the end of the function.